Add files via upload
This repository contains the web scrapping assignment-3. I have scraped various types of data like images and tables for 7 different questions given in the assignment. I have used different libraries for the scraping and Pandas library to make data frames and save results in the form of a CSV file.

Please find the jupyter notebook namely 'Web Scrapping Assignment 3.ipyb' in the repository.